Don't slash finalizer when double-vote has invalid signature
Resolves #680

The issue was that we recored all votes, even if they where invalid.
Then when a valid vote arrives, we count it as a double-vote
and create slash transaction.

Also, in this PR I've rewritten the esperanza_slash.py as the
previous one was broken. Currently, we test only double-votes.
In the follow-up PR I'll add tests for surrounded votes.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>