Replace bitcoin coinbase with PoS coinbase (#738)
This PR replaces the coinbase from bitcoin with a PoS coinbase. We then connect the staking mechanism (validation is not completely wired in yet, it will be with #577). Most changes are test related since this is such a radical change. 
The main TODO left to be solved is to completely get rid of the `miner.cpp` and replace it with the logic in `proposer.cpp`.

To summarize the changes made by this PR:
- Introducing a new PoS coinbase, replacing the old bitcoin one. This coinbase needs one (or more) 
stake inputs and has at least two outputs (the reward and the stake out). If the stake is being split 
then we expect multiple outputs.
- Now the coinbase maturity is applied only to the first output (the other outs are the stake that is already mature by definition).
- Adapted part of the validation logic to account for the fact that the coinbase transaction now spends tokens.
- The coinbase now is necessarily a segwit transaction.
- Adapting  the vast majority of the functional tests to the new PoS coinbase.

Co-authored-by: Mihai Ciumeica <mihai@thirdhash.com>
Co-authored-by: Julian Fleischer <julian@thirdhash.com>
Co-authored-by: Aleksandr Mikhailov <aleksandr@thirdhash.com>
Co-authored-by: Mateusz Morusiewicz <mateusz@thirdhash.com>
Co-authored-by: Azat Nizametdinov <azat@thirdhash.com>
Co-authored-by: Matteo Sumberaz <matteo@thirdhash.com>

Related: #637 

Signed-off-by: Matteo Sumberaz <matteo@thirdhash.com>
Signed-off-by: Mateusz Morusiewicz <mateusz@thirdhash.com>
Signed-off-by: Julian Fleischer <julian@thirdhash.com>
Signed-off-by: Mateusz Morusiewicz <mateusz@thirdhash.com>
Signed-off-by: Aleksandr Mikhailov <aleksandr@thirdhash.com>
Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>
Signed-off-by: Mateusz Morusiewicz <mateusz@thirdhash.com>
Signed-off-by: Azat Nizametdinov <azat@thirdhash.com>