Don't set epochLength=5 in regtest as it's a default value
The reason of dropping this parameter is to clearly show
which parameters are required to be changed.

In the follow up PR I'll adjust other tests to default epochLength=5
where it's not required to have a longer one. The reason is
to reduce the debug.log in case the test fails.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>