Rename check_finalization to assert_finalizationstate
Reasons of renaming:
1. test_framework uses assert_* instead of check_* prefix
for functions that crash the test when comparison of objects fails.
For me personally, check_* looks like a function that returns bool value
but would be acceptable if it was the framework naming convention.
Examples: assert_fee_amount, assert_raises_rpc_error
2. use finalizationstate instead of finalization to align with the
RPC/object naming.

The function is moved to the "Assert functions" block.

Also, I've fixed the error message:
before:
```
AssertionError: not(2 == 3)
```
now:
```
AssertionError: currentEpoch: not(6 == 7)
```

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>