Move witness commitment from coinbase to header (#798)
This moves the witness commit from the coinbase to the header.

In bitcoin, as they could not change the block structure without hardforking, the witness commitment is added as an unspendable additional output to the coinbase transaction. That witness commitment contains the witness nonce (to distinguish it from other random junk that people append to the coinbase) and the witness merkle tree root.

In unit-e, as in particl, we move that witness merkle tree root (hash) to the header. This is done so that we can include the coinbase wtxid in it. Otherwise the wtxid (which would also cover the witness merkle root as it encompasses the whole transaction including witnesses) would have to... hash itself?!

We need it in order to properly sign the staking inputs. Also we have a very old uip for it: https://github.com/dtr-org/uips/blob/master/UIP-0003.md

Other things in this pull requests:

- adds templates for `util::to_string` for `std::unique_ptr`, `std::shared_ptr`, and regular pointers such that `util::to_string` works e.g. for `std::vector<CTransactionRef>` which comes in quite handy in tests and log statements
- fixes `assert_not_equal` which was broken before
- fixes an issue in `net_processing` which made `p2p_compactblocks` flaky (to be backported to bitcoin)

Rebased version of https://github.com/dtr-org/unit-e/pull/756

Signed-off-by: Julian Fleischer <julian@thirdhash.com>