Add finalizer commits merkle root to block header
This PR adds merkle root of finalizer commits to the block header.
Having this merkle root in a header will allow us to verify
the consistency of headers+commits P2P messages.

Notice, original "hashMerkleRoot" also takes finalizer commits into account.
The reason is that it's needed for Bloom Filter. We can review
the possibility of not including finalizer commits into hashMerkleRoot
and refactor Bloom Filters/SPV clients in a separate issue or PR.

We also have an issue how to optimize our merkle roots https://github.com/dtr-org/unit-e/issues/806
and this will be discussed/tackled in a separate PR.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>