Implement proper finalization state repository saving and loading (leveldb) (#793)
Introduces finalization::StateDB component which is basically an interface to load and save repository. It provides two ways to restore repository: 
* `StateDB::Load` loads every entry from the database to the repository.
* `StateDB::LoadStatesHigherThan` loads states for the blocks which height is greater than given height.

The second one is preferable as we keep only a limited set of states in memory. In other words, even when we load everything from the database, we trim the repository on the next step.

Also implements recovery for cases when finalization state database is partially broken or out-of-sync with the block-index database.

Adds `BlockIndexMap` component which is a mockable abstraction for `mapBlockIndex`.

Fixes #432.

Signed-off-by: Stanislav Frolov <stanislav@thirdhash.com>