Add keytools to functional tests (#842)
Add KeyTool to functional tests

KeyTool is a utility for dealing with key management. It wraps RPC calls
and CECKey such that keys can be generated in python or on the node
and uploaded/downloaded to/from the node. It is capable of generating
bech32 addresses, legacy addresses, and lookup private/public keys
by address from the node, as well as deriving public keys from any given
private key, both compressed and uncompressed.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>