Fixes in esperanza protocol for logout and withdraw txs
This PR makes several fixes for Esperanza protocol, more particularly for logout and withdraw logic
as it currently can lead to stuck of finalization.

1. Retrospective bug of  epoch=1 PR https://github.com/dtr-org/unit-e/pull/809.
We shouldn't increment the dynasty for finalized hardcoded epoch=0 as it's already considered incremented from -1.
In ideal scenario, when we always reach finalization, dynasty number catches up to finalization but not another way around.
Otherwise, it will lead to stuck of finalization. `rpc_finalization.py` test was improved to catch it.
2. int64_t-> uint64_t bug in `GetDynastyDelta`. This function returns the delta that is added to `m_cur_dyn_deposits`
during dynasty increment. If finalizer logged out, `GetDynastyDelta` returns `-deposit` value which is added to `m_cur_dyn_deposits`
to decrease it but since the returned type of `GetDynastyDelta` was `uint64_t`,
we never decreased total deposits and it means that we couldn't reach 2/3 to justify the epoch.
3. Bug in `DepositExists`. It returned `true` only if we had deposits in the current dynasty
and in the previous one. The issue was that once we increment the dynasty and register deposits,
finalizers can start voting but their votes are rejected as node considers that we shouldn't have votes yet.
The reason we didn't spot this problem in functional tests is that we thought finalization happened due to votes
but actually, it happened due to InstaJustify.  Now `DepositExists` returns `true` only if current dynasty has deposits.
`rpc_finalization.py` is adjusted to spot this problem too.
4. Bug in how 2/3 of votes are calculated.
We checked if we reached 2/3 by comparing current votes with the threshold in the current dynasty and in the previous one.
The issue is if someone logged out, you'll have a point where threshold in dynasty X is smaller then in dynasty X-1
and it means that if 1/3 of deposits are withdrawn, you can't pass the check `current_votes >= prev_dynasty_threahold`.
The fix is to check if we reached 2/3 of votes for the current dynasty only.

All the above bugs where found by updated `esperanza_withdraw.py` test.
This test and `rpc_finalization.py` can be a good starting point to understanding changes.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>