Allow unit tests being run in parallel again (#870)
Currently unit tests can not be run in parallel. This prevents us from speeding up ci builds (see https://github.com/bitcoin/bitcoin/pull/12926, https://github.com/dtr-org/unit-e/pull/865) and won't work with the updated build definitions from bitcoin 0.17 (#860). This is a regression which was introduced in https://github.com/dtr-org/unit-e/pull/793. This pull request fixes https://github.com/dtr-org/unit-e/issues/861.

The problem is that the `StateDB` component does something actively when being initialized (not something a component should do), which is to access disk and so on, which can fail. Since there is a `UnitEInjector` for every `BasicTestingSetup` running the unit tests in parallel creates a database _in the same directory_ per unit tests suite which clashes and breaks and throws and aborts and dies. This can be prevented by using an in-memory database, but the problem shifts: How to get that configuration in?

I did not want to expose this as a user-definable setting and also not as a chain parameter, so these two configuration options are not available. Which is why I created `UnitEInjectorConfiguration`. This is something which should not be necessary as usually you would not expose the same module as in production in unit tests but use a subset or just the mocked version, but since the rest of bitcoin is not in well-defined components and accesses them using `GetComponent`  a global injector has to be available in some tests.

The change ultimately does not affect the prod version of things, but the tests, which inject a different `UnitEInjectorConfiguration`. I made this a `struct` rather then just a flag `unit_tests` such that it is extensible in case we have other cases like this. In order to make the injector be able to access fields from it's own instance in the definition of a component I altered the definition of `UNMANAGED_COMPONENT` a bit.

I would like to point out that these initialization issues would also have happened with an Application class (#723), as that is exactly what `UnitEInjector` is, just you would inject things manually.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>