Add getrawsnapshot RPC
In functional tests we use p2p interface to retrieve the snapshot
before broadcasting to other nodes. It works fine but now we want to
extend tests and broadcast the valid snapshot but which is not
part of finalized epoch. The issue is that from p2p we can't
receive the non-finalized snapshot to simulate this scenario.

Having an RPC command to retrieve any available snapshot
will help us to extend fast sync tests.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>