Restore wallet txs after the fast sync
To be able to restore wallet transactions, before syncing
the node must be configured appropriately:
1. node must import the private key
2. node must generate addresses up to(including) the last one that was used.
3. (Alternatively), we can pre-generate few K addresses (as the rescan does)

Pre-generating addresses is required as since the snapshot
doesn't contain the full blockchain history, node can't recover all its addresses.

After the sync, we have two limitations:
1. since the snapshot doesn't contain inputs, txs that
end up in the wallet, won't have inputs too. It shouldn't
affect the wallet but will require us to tune the GUI
and properly visualize it. Alternatively, we can request
full transaction from the peers on demand.
2. `CMerkleTx` has a field `nIndex` which contains the position
of the transaction in the block. We also don't have this field
and it's set to 0. Would be nice to understand what is the impact
of it. If there is a need, we can also download a full block on request.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>