Configure on which block of the epoch a finalizer starts voting
This PR introduces `-finalizervotefromepochblocknumber` parameter that
allows to configure from which block of the epoch finalizer can start voting.

It can make sense to start voting a bit later than after the checkpoint
is processed as if re-orgs happens on a checkpoint and finalizer switches to
that fork, it can't vote on this epoch again as it will be double-voting.
Delaying the vote reduces the chance of voting on the fork
that won't be final for the finalizer but delaying for too long
can risk that vote won't be included by proposer at all.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>