Adjust logging in finalization state processor (#912)
During testnet we figured out that scenario when peer sends
outdated commits/blocks is a normal case scenario, and in fact
too often one. When node syncs with two peers, it receives commits
and blocks from first one. It leads to trimming finalization state
repository. Then the node starts receving commits from second
peer with some delay and prints dozen of errors which actually is
not an error but just a fact that full sync completed with another
peer.

I'm keeping the log message because it might be helpful in debugging,
but removing misleading "ERROR".

Signed-off-by: Stanislav Frolov <stanislav@thirdhash.com>