Remove PoW (#929)
The commits, in order:
1. Removes PoW and moves `(Contextual)?CheckBlock(Header)?` functions to a component called `LegacyValidationInterface`
2. Removes now useless `nonce` from header

The `LegacyValidationInterface` is a proper component and has two implementations: One that harbours the current `*Check*` functions (moved them, only thing I did is I adapted the code style as it's added in a new compilation unit). Unit tests and functional tests pass exactly as before with it.

A new implementation, since the API of `BlockValidator` and `StakeValidator` are different, will translate between the error-codes returned by `BlockValidator::*Check*` and `CValidationState`. In this pull this implementation is left empty for now to reduce the size of the changeset.

Reincarnation of https://github.com/dtr-org/unit-e/pull/433

Signed-off-by: Julian Fleischer <julian@thirdhash.com>