Fix coinbase transaction amount returned by filtertransactions
Currently, a transaction amount and fee are calculated by the
following algorithm:

* The 'fee' is the difference between a transaction's sum of outputs and
its sum of inputs.

* The 'amount' is the difference between the sum of its outputs and the
sum of those outputs that are sent to the node itself.

This makes sense for a regular transaction, however, for a coinbase Tx,
the amount will be zero, because all of its outputs are sent to the node
itself.

This commit special-cases coinbase Txs and sets the amount to the fee
itself, which is what most people expect to see.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>