Calculate and store stake_modifier, Fix functional tests (#974)
The stake_modifier for a block is saved in `CBlockIndex` and it was not set before. This pull request changes that.

Updated the functional tests such that the snapshot hash calculation correctly takes `stake_modifier` into account. Notable change: Instead of passing `snapshot_meta.hash` it passes `snapshot_meta` which harbours both `hash` and `stake_modifier`.

Signed-off-by: Julian Fleischer <julian@thirdhash.com>