Fix inconsistent ValidatorState after reorgs
Resolves https://github.com/dtr-org/unit-e/issues/972

It's the follow-up commit of https://github.com/dtr-org/unit-e/pull/937

In this commit we instead of caching the finalizer's phase in `ValidatorState.m_phase`,
we compute it on demand in `ValidatorState::Phase WalletExtension::GetFinalizerPhase`.
This resolves the issue of unsynchronization `ValidatorState` with `FinalizerState::Validator` during re-orgs.

Also,` ValidatorState::Phase` was extended to show more precisely in which phase the finalizer is.
The full list is:

* NOT_VALIDATING
* WAITING_DEPOSIT_CONFIRMATION
* WAITING_DEPOSIT_FINALIZATION
* IS_VALIDATING
* WAITING_FOR_WITHDRAW_DELAY
* WAITING_TO_WITHDRAW

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>