Change TxType type from uint16_t to uint8_t
This type is used in `Coin` and `snapshot::UTXOSubset` and we serialize
it as `uint8_t` to reduce the disk usage.

Since `TxType` is `uint16_t` there is a risk that we create a record
that will have a number larger than 255 and it will cause incorrect
value to be stored in the Coin or transferred via snapshot.

To make it consistent on how this field is used,
we should stick to `uint8_t` everywhere.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>