Workaround for issue 962
Although the ideal would be to remove the "deprecated" coinbase
transaction from the wallet, we don't do that because it's not yet clear
what should we do with other transactions that use as input the outputs
of the former one.

So for now I opted for changing the log message in order to distinguish
the coibase case, which will be far more common than the case where the
user manually generates conflicting transactions after a re-org.

Signed-off-by: Andres Correa Casablanca <andres@thirdhash.com>