Show failing unit tests at the bottom in run-unit-tests.sh
This commit modifies the sort order for the results of run-unit-tests.sh.

Previously, they were sorted lexicographically, which meant that the
failing tests showed on top and could potentially disappear off-screen.
Now, the test results are sorted first by name (from character 6 to the
end of the line), then by status (characters 3 to 5) in reverse order.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>