Fix incorrect mapping between dynasty and epoch
The bug was that epoch number in `m_dynasty_start_epoch` was one less than it should be.
This allows finalizer to withdraw one epoch earlier than expected.
Test `rpc_finalization.py` was extended to catch the bug and `esperanza_widthdraw.py` was
adjusted accordingly.

The `m_epoch_to_dynasty` mapping was removed (instead of fixing it) as it wasn't used.
Supporting fewer states reduces the chance of data unsynchronization.

Signed-off-by: Kostiantyn Stepaniuk <kostia@thirdhash.com>