Fix mempool_persist.py
Bitcoin 0.17 introduced a global flag `g_is_mempool_loaded`, which is
set after the import thread finishes, and is checked when saving the
mempool (because we can't save it if we haven't loaded it yet).

This commit adds the same check to `getrawmempool`, so that the functional
tests can at least wait until the flag is set before trying to act on it.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>