Validate deposit prior adding to the wallet (#1040)
At the moment, when a user prepares deposit, the node creates a transaction and adds it to the
wallet.  It, in turn, tries to accept it to the memory pool and could fail when the deposit is
invalid (for example validator is blacklisted). That leads to disappearing of the deposit from the
balance.

This commit fixes that by validating deposit before adding transaction to the wallet.

Fixes #996.

Signed-off-by: Stanislav Frolov <stanislav@thirdhash.com>