Reduce Travis log spam
Previously, each Makefile script which runs each individual unit test was
displayed for debugging purposes. This commit suppresses this output,
which should make the Travis log output smaller.

Note that, just after the unit tests, we also test `unit-e-tx`, which also
outputs its testcases. This output should not be suppressed, since the
test takes a long time, so the Travis build can timeout if it doesn't
see any output for a long time.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>