Fix scripted diffs with CRs in the commit message
GitHub inserts line endings using CRs in some cases when squashing
commits. This causes issues in case of scripted diffs where the
script is extracted from the commit message by `commit-script-check.sh`.

This patch fixes this by removing CRs at the end of the line before
extracting the script for checking the commit.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>