Run all lint checks
Include lint checks which were added in 0.18 in the lint-all script
so they are all run.

Also fix the logic to be the one from the unit-e
master branch to always run all checks and not stop on the first
failure and remove the obsolete LINTALL variable.

Some of the checks fail on the upstream sync branch. We'll need to
fix them. I'll add todos to the pull request.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>